requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
django-tinymce==3.6.1
Pillow==10.0.1
sqlparse==0.4.4
